Register finalizado :)

+++ Segundo Año/Programacion_Frontend/TP-backendFront/api-users/controllers/controladorUsuario.js
@@ -1,19 +1,23 @@
 const Usuario = require("../models/modelUsuario");
+const TokenCheck = require("../middlewares/verificacionJwt");
 
 // Obtiene los datos de todos los usuarios
 exports.obtenerUsuarios = async (req,res) => {
-    try { // Intenta realizar la consulta con el metodo findAll
-        const usuarios = await Usuario.findAll()
-        res.status(200).json({ // Si tuvo exito, devuelve todos los datos
-            estado:"Ok",
-            usuarios
-        })
-    } catch (error) { // Si no pudo hacerlo, devuelve un error 500 y un mensaje de error
-        console.error(error);
-        res.status(500).json({
-            estado:"Error",
-            mensaje:"Error del servidor"
-        })
+    const validado = TokenCheck.verificacion(req,res);
+    if (validado == true) {
+        try { // Intenta realizar la consulta con el metodo findAll
+            const usuarios = await Usuario.findAll()
+            res.status(200).json({ // Si tuvo exito, devuelve todos los datos
+                estado:"Ok",
+                usuarios
+            })
+        } catch (error) { // Si no pudo hacerlo, devuelve un error 500 y un mensaje de error
+            console.error(error);
+            res.status(500).json({
+                estado:"Error",
+                mensaje:"Error del servidor"
+            })
+        }
     }
 }
 
@@ -34,7 +38,7 @@ exports.crearUsuario = async(req,res) => {
                 mensaje:`Usuario ${usuario.username} creado con exito!`
             })
         } else {
-            res.status(200).json({
+            res.status(400).json({
                 estado:"Error",
                 mensaje:`El nombre de usuario ${usuarioEntrante} ya se encuentra ocupado, ingrese otro`
             })
@@ -79,48 +83,54 @@ exports.obtenerUsuarioUnico = async (req,res) => {
 
 // Borrar usuario
 exports.borrarUsuario = async (req,res) => {
-    try { // Borra un usuario con el metodo destroy, buscando el usuario que coincida con el id enviado
-        const usuario = await Usuario.destroy({
-            where: {
-                id: req.params.id
-            }
-        })
-        res.status(200).json({ // Si lo pudo eliminar, devuleve un ok y el mensaje de operación exitosa
-            estado:"Ok",
-            message:`Usuario ${usuario.username} eliminado`
-        })
-    } catch(error) { // Si falló devuelve error en "estado" y los detalles del mismo por consola
-        console.error(error);
-        res.status(500).json({
-            estado:"Error",
-            mensaje:"Ocurrió un error al eliminar el usuario"
-        })
+    const validado = TokenCheck.verificacion(req,res);
+    if (validado == true) {
+        try { // Borra un usuario con el metodo destroy, buscando el usuario que coincida con el id enviado
+            const usuario = await Usuario.destroy({
+                where: {
+                    id: req.params.id
+                }
+            })
+            res.status(200).json({ // Si lo pudo eliminar, devuleve un ok y el mensaje de operación exitosa
+                estado:"Ok",
+                message:`Usuario ${usuario.username} eliminado`
+            })
+        } catch(error) { // Si falló devuelve error en "estado" y los detalles del mismo por consola
+            console.error(error);
+            res.status(500).json({
+                estado:"Error",
+                mensaje:"Ocurrió un error al eliminar el usuario"
+            })
+        }
     }
 }
 
 //Actualizar usuario
 exports.actulizarUsuario = async (req,res) => {
-    // Crea las contantes con el id y los datos actualizados...
-    const userId = req.params.id;
-    const dataUsuario = req.body;
-    // Para luego crear el objeto usuario listo para lanzar con el update
-    const usuario = { userId, ...dataUsuario }
-    try { // Utiliza el metodo update con la informacion actualizada donde el id coincida con el del usuario que se queiere corregir
-        const actualizacionUsuario = await Usuario.update({ ...usuario }, {
-            where: { 
-                id:userId
-            }
-        })
-        res.status(200).json({
-            estado:"OK",
-            mensaje:"Usuario actualizado"
-        })
-    } catch (error) { // Igual que las consultas anteriores, si falla retorna el error 500 y el mensaje por terminal
-        console.error(error);
-        res.status(500).json({
-            estado:"Error",
-            mensaje:"Error al actualizar el usuario"
-        })
+    const validado = TokenCheck.verificacion(req,res);
+    if (validado == true) {
+        // Crea las contantes con el id y los datos actualizados...
+        const userId = req.params.id;
+        const dataUsuario = req.body;
+        // Para luego crear el objeto usuario listo para lanzar con el update
+        const usuario = { userId, ...dataUsuario }
+        try { // Utiliza el metodo update con la informacion actualizada donde el id coincida con el del usuario que se queiere corregir
+            await Usuario.update({ ...usuario }, {
+                where: { 
+                    id:userId
+                }
+            })
+            res.status(200).json({
+                estado:"OK",
+                mensaje:"Usuario actualizado"
+            })
+        } catch (error) { // Igual que las consultas anteriores, si falla retorna el error 500 y el mensaje por terminal
+            console.error(error);
+            res.status(500).json({
+                estado:"Error",
+                mensaje:"Error al actualizar el usuario"
+            })
+        }   
     }
 }
